refactor(shell): extract spawn func into its own file
